requirements.txt
selenium==4.16.0
pytest==7.4.4
webdriver-manager
python-dotenv==1.0.1